requirements.txt
pillow==11.2.1
lxml==5.4.0
